+++ modules/openapi-generator/src/main/java/org/openapitools/codegen/CodegenResponse.java
@@ -614,4 +614,9 @@ public class CodegenResponse implements IJsonSchemaValidationProperties {
 
     @Override
     public void setHasMultipleTypes(boolean hasMultipleTypes) { this.hasMultipleTypes = hasMultipleTypes; }
+
+    @Override
+    public String getBaseType() {
+        return baseType;
+    }
 }
