+++ modules/openapi-generator/src/main/java/org/openapitools/codegen/IJsonSchemaValidationProperties.java
@@ -1,6 +1,10 @@
 package org.openapitools.codegen;
 
+import java.util.Collections;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Stream;
 
 import io.swagger.v3.oas.models.media.Schema;
 import org.openapitools.codegen.utils.ModelUtils;
@@ -212,5 +216,45 @@ public interface IJsonSchemaValidationProperties {
         } else if (ModelUtils.isAnyType(p)) {
             setIsAnyType(true);
         }
+    }
+
+    /**
+     * @return basic type - no generics supported.
+     */
+    default String getBaseType() {
+        return null;
+    };
+
+    /**
+     * @return complex type that can contain type parameters - like {@code List<Items>} for Java
+     */
+    default String getComplexType() {
+        return getBaseType();
+    };
+
+    /**
+     * Recursively collect all necessary imports to include so that the type may be resolved.
+     *
+     * @param includeContainerTypes whether or not to include the container types in the returned imports.
+     * @return all of the imports
+     */
+    default Set<String> getImports(boolean includeContainerTypes) {
+        Set<String> imports = new HashSet<>();
+        if (includeContainerTypes || !(this.getIsArray() || this.getIsMap())) {
+            if (this.getComposedSchemas() != null) {
+                CodegenComposedSchemas composed = (CodegenComposedSchemas) this.getComposedSchemas();
+                List<CodegenProperty> allOfs =  composed.getAllOf() == null ? Collections.emptyList() : composed.getAllOf();
+                List<CodegenProperty> oneOfs =  composed.getOneOf() == null ? Collections.emptyList() : composed.getOneOf();
+                Stream<CodegenProperty> innerTypes = Stream.concat(allOfs.stream(), oneOfs.stream());
+                innerTypes.flatMap(cp -> cp.getImports(includeContainerTypes).stream()).forEach(s -> imports.add(s));
+            } else {
+                imports.add(this.getComplexType());
+                imports.add(this.getBaseType());
+            }
+        }
+        if (this.getItems() !=null) {
+            imports.addAll(this.getItems().getImports(includeContainerTypes));
+        }
+        return imports;
     }
 }
